Minimize amount of code needed to define locale
Also minimize code duplication by defining abstract class provider, which can be implemeted and overriden if needed
